Splitting monolith mock header to speed up test compilation
