+++ test/mocks/server/config_tracker.cc
@@ -1,0 +1,37 @@
+#include "config_tracker.h"
+
+#include <string>
+
+#include "envoy/admin/v3/server_info.pb.h"
+#include "envoy/config/core/v3/base.pb.h"
+
+#include "common/singleton/manager_impl.h"
+
+#include "gmock/gmock.h"
+#include "gtest/gtest.h"
+
+using testing::_;
+using testing::Invoke;
+using testing::Return;
+using testing::ReturnPointee;
+using testing::ReturnRef;
+using testing::SaveArg;
+
+namespace Envoy {
+namespace Server {
+MockConfigTracker::MockConfigTracker() {
+  ON_CALL(*this, add_(_, _))
+      .WillByDefault(Invoke([this](const std::string& key, Cb callback) -> EntryOwner* {
+        EXPECT_TRUE(config_tracker_callbacks_.find(key) == config_tracker_callbacks_.end());
+        config_tracker_callbacks_[key] = callback;
+        return new MockEntryOwner();
+      }));
+}
+
+MockConfigTracker::~MockConfigTracker() = default;
+
+
+
+}
+
+}
